Retain port binding profile after live migration
This patch addresses an issue where the binding
profile of a port is lost after live-migration.

This patch fixes it by correctly populating the
profile field in the
`check_can_live_migrate_destination` method of
nova-compute. This ensures the existing binding
profile of the port is retained during the port
binding in nova-conductor.

Closes-Bug: #2089566
Change-Id: I93dae9e9a1866e62efb9aa070e08619e7d44c1a2